Rename metric evaluator modules
Use clearer metric_evaluator and noise_clamper module names so in-tree imports match the utilities' roles without changing their public symbols.

Made-with: Cursor
