+++ projects/rocprofiler-compute/src/utils/metrics/noise_clamper.py
